Changes to the whois command
Fetching messages at startup, mostly, and improvements to embed creation.
